pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "memory_architect"
version = "0.1.0"
description = "Adaptive Policy-Controlled Memory Governance System"
authors = []

[tool.poetry.dependencies]
python = "^3.13"
pydantic = "^2.6.0"        # For strict schema validation
pydantic-settings = "^2.1.0" # For YAML config management [6]
langchain = "^0.1.0"       # Orchestration
networkx = "^3.2.1"          # Graph memory overlay
presidio-analyzer = "^2.2.354" # PII Detection [8]
presidio-anonymizer = "^2.2.354" # PII Redaction
spacy = "^3.7.0"
sentence-transformers = "^2.3.0" # Local embeddings
numpy = "^1.26"
PyYAML = "^6.0"            # For parsing policy files
llama-cpp-python = "^0.3.16"
fastapi = "^0.124.2"
uvicorn = "^0.38.0"
httpx = "^0.28.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
black = "^24.0"
isort = "^5.13"
